+++ main.c
@@ -4,132 +4,8 @@
 #include "SDL.h"
 #include "SDL_image.h"
 #include "main.h"
-
-#define GRAVITY 0.8f
-
-int processEvent(SDL_Window* window, GameState* game)
-{
-    SDL_Event event;
-    int done = 0;
-
-    while (SDL_PollEvent(&event)) {
-        switch (event.key.keysym.sym) {
-            case SDLK_ESCAPE:
-                done = 1;
-                break;
-                //case SDLK_UP:
-                //	if(game->man.dy == 0.0f)
-                //	{
-                //	game->man.dy = -15;
-                //	}
-                //break;
-        }
-    }
-
-    const Uint8* state = SDL_GetKeyboardState(NULL);
-    if (state[SDL_SCANCODE_ESCAPE]) {
-        done = 1;
-    }
-    if (state[SDL_SCANCODE_LEFT]) {
-        game->man.dx -= 0.6f;
-        game->man.facingLeft = 1;
-        if (game->man.dx <= -6.5f) {
-            game->man.dx = -6.5f;
-        }
-    }
-    if (state[SDL_SCANCODE_RIGHT]) {
-        game->man.dx += 0.6f;
-        game->man.facingLeft = 0;
-        if (game->man.dx >= 6.5f) {
-            game->man.dx = 6.5f;
-        }
-    }
-
-    if (state[SDL_SCANCODE_RIGHT] == state[SDL_SCANCODE_LEFT]) {
-        game->man.dx *= 0.8f;
-    }
-    /*
-	if(state[SDL_SCANCODE_DOWN])
-	{
-		game->man.y += 5;
-	}
-	*/
-    if (state[SDL_SCANCODE_UP]) {
-        game->man.dy = -15;
-    }
-
-    //scrolling
-
-    if (game->man.x + game->man.w > game->rightScreen) {
-        game->rightScreen = game->man.x + game->man.w;
-        game->leftScreen = game->rightScreen - 920;
-
-        game->scrollX = -(game->rightScreen - 1000);
-    }
-    if (game->man.x < game->leftScreen) {
-        game->leftScreen = game->man.x;
-        game->rightScreen = game->leftScreen + 920;
-
-        game->scrollX = -(game->leftScreen - 80);
-    }
-    if (game->man.y < game->topScreen) {
-        game->topScreen = game->man.y;
-        game->bottomScreen = game->topScreen + 560;
-
-        game->scrollY = -(game->topScreen - 80);
-    }
-    if (game->man.y + game->man.h > game->bottomScreen) {
-        game->bottomScreen = game->man.y + game->man.h;
-        game->topScreen = game->bottomScreen - 560;
-
-        game->scrollY = -(game->bottomScreen - 560);
-    }
-    return done;
-}
-
-void collisionDetection(GameState* game)
-{
-    for (int i = 0; i < 100; ++i) {
-        float mw = game->man.w, mh = game->man.h;
-        float mx = game->man.x, my = game->man.y;
-        float lx = game->ledges[i].x, ly = game->ledges[i].y, lw = game->ledges[i].w, lh = game->ledges[i].h;
-        float vCheck = fabsf((my + mh / 2) - (ly + lh / 2));
-
-        if (mx + mw > lx && mx < lx + lw && vCheck > 5) {
-            if (game->man.dy > 0 && my + mh > ly && my < ly) {
-                game->man.y = ly - mh;
-                my = ly - mh;
-                game->man.dy = 0;
-            }
-
-            if (game->man.dy < 0 && my < ly + lh && ly + lh < my + mh) {
-                game->man.y = ly + lh;
-                my = ly + lh;
-                game->man.dy = 0;
-            }
-        }
-        if (my + mh > ly && my < ly + lh) {
-            if (game->man.dx > 0 && mx + mw > lx && mx < lx) {
-                game->man.x = lx - mw;
-                mx = lx - mw;
-                game->man.dx = 0;
-            }
-
-            if (game->man.dx < 0 && mx < lx + lw && lx + lw < mx + mw) {
-                game->man.x = lx + lw;
-                mx = lx + lw;
-                game->man.dx = 0;
-            }
-        }
-    }
-}
-
-void processGravity(GameState* game)
-{
-    game->man.y += game->man.dy;
-    game->man.dy += GRAVITY;
-    game->man.x += game->man.dx;
-}
+#include "event.h"
+#include "physics.h"
 
 void GameInit(GameState* game)
 {
